==> notebooks/01_load_and_validate_raw_data.ipynb <==
# Downloading Dataset from NYC Trip Data
This website provides the historical of taxi data in NYC. The website can be accessed through this url https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page

from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """
    Funtion to download data from NYC Trip Data
    """
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'  # month:02d -> 2 digit value start by 0 (format for month)
    response = requests.get(URL)

    if response.status_code == 200: # status ok: request was fulfilled
        path = f'../data/raw/rides_{year}--{month:02d}.parquet'
        open(path, "wb").write(response.content)  # wb = writing binary file
        return path
    else:
        raise Exception(f'{URL} is not available')


download_one_file_of_raw_data(year=2023, month=1)

import pandas as pd
rides = pd.read_parquet('../data/raw/rides_2023--01.parquet')

rides.head(5)

# Selecting the required columns
rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.head(5)

# Renaming the column

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id'},
    inplace = True
)

rides.head(5)

rides['pickup_datetime'].describe()

# Selecting data from 1st january - 31 january 2023
rides = rides[rides.pickup_datetime >= '2023-01-01']
rides = rides[rides.pickup_datetime < '2023-02-01']
rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2023_01.parquet')
